pkvm: implement the mmio emulation via the hypervisor
As an alternative MMIO emulation mode implement the MMIO
emulation in the hypervisor. Ultimately the #VE emulation
mode is more suitable for TDX based systems where even the
hypervisor need not be trusted, but this is not really
the case with the PKVM model. The hypervisor remains as a
fundamental piece of our TCB.

Benefits of the model are that it works on every X86 HW
flavor (including mobile) and it doesn't require custom
guest BIOSs. Any BIOS will work as the guest SMM emulation
is executed by the hypervisor and not by the untrusted host.
Moreover, the guest can have a functional, somewhat secure
SMM and use it.

While the changes now seem working, the work here is not
complete. Several additional hardening steps are required,
especially as the hypervisor is using the VMM mm_struct to
make sense of the userspace addresses.

Signed-off-by: Janne Karhunen <Janne.Karhunen@gmail.com>